Extract InitFreeImage
